pyproject.toml
[tool.poetry]
name = "contentful"
version = "2.1.1"
description = "Contentful SDK for the Content Delivery API"
authors = ["Contentful GmbH <bhushan.lodha@external.contentful.com>"]
license = "MIT"
readme = "README.rst"
repository = "https://github.com/contentful/contentful.py"
documentation = "https://contentful.github.io/contentful.py/"
packages = [{ include = "contentful" }]
include = ["CHANGELOG.md"]
keywords = [
    "contentful",
    "delivery",
    "cda",
    "cms",
    "content",
]
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Topic :: Software Development :: Libraries',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Natural Language :: English',
    'Programming Language :: Python :: 2',
    'Programming Language :: Python :: 2.7',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.1',
    'Programming Language :: Python :: 3.2',
    'Programming Language :: Python :: 3.3',
    'Programming Language :: Python :: 3.4',
    'Programming Language :: Python :: 3.5',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
python-dateutil = "^2.8"
requests = "^2.20"
aiohttp = "^3.9"
orjson = { version = "^3.9", markers = "platform_python_implementation != 'PyPy'" }

[tool.poetry.group.test.dependencies]
vcrpy = "^6"
coverage = "^7"
flake8 = "^7"
tox = "^4"
requests-mock = "^1.5"
aioresponses = "^0.7"

[tool.poetry.group.docs.dependencies]
Sphinx = "^7"
sphinxcontrib-websupport = "^1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
